+++ ica-rs/ica_typing.py
@@ -2,7 +2,6 @@
 
 from typing import Callable, Tuple
 
-
 """
 ica.rs
 pub type RoomId = i64;
@@ -78,6 +77,13 @@ class IcaSendMessage:
         """
         self.content = content
         return self
+    def set_img(self, file: bytes, file_type: str, as_sticker: bool):
+        """
+        设置消息的图片
+        @param file: 图片文件 (实际上是 vec<u8>)
+        @param file_type: 图片类型 (MIME) (image/png; image/jpeg)
+        @param as_sticker: 是否作为贴纸发送
+        """
 
 
 class IcaDeleteMessage: