+++ ica-rs/src/py/class/ica.rs
@@ -141,6 +141,10 @@ impl SendMessagePy {
     pub fn get_content(&self) -> String { self.msg.content.clone() }
     #[setter]
     pub fn set_content(&mut self, content: String) { self.msg.content = content; }
+    /// 设置消息图片
+    pub fn set_img(&mut self, file: Vec<u8>, file_type: String, as_sticker: bool) {
+        self.msg.set_img(&file, &file_type, as_sticker);
+    }
 }
 
 impl SendMessagePy {